fix: ID of entity.User should be int64 instead of int

+++ services/auth/internal/controller/http/middleware/auth.go
@@ -18,7 +18,7 @@ const (
 
 type Claims struct {
 	jwt.RegisteredClaims
-	UserID int    `json:"id"`
+	UserID int64  `json:"id"`
 	Login  string `json:"login"`
 }
